+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 ## Release Notes
 
+### 1.12.3
+* Integrated SCM API (best user experience when `vscode-clearcase.useClearDlg` is set to false)
+* Untracked files are listed in the SCM View, file types can be configured by `vscode-clearcase.viewPrivateFileSuffixes`
+* Checkedout files are listed in the SCM View
+* Checkin all files via the SCM View, also add comment
+* New context menu `delete file` for view private files
+
 ### 1.12.2
 * Updated to last vscode dependency to avoid url-parse vulnerability